Update akka remoting test and move to dev subproject.
This commit moves the RemoteClient and RemoteServer -- a mini-demo for
development testing -- to a "dev" subproject for development/experimental
code.
