+++ dev/src/main/resources/application.conf
@@ -2,7 +2,7 @@ remoteServer {
   include "common"
 
   akka {
-    remote.server.port = 2552
+    remote.netty.port = 8552
     remote.server.host = "0.0.0.0"
     remote.server.message-frame-size = 16777216
     remote.client.message-frame-size = 16777216
@@ -12,8 +12,10 @@ remoteServer {
 remoteClient {
   include "common"
 
+  geotrellis.clusterSeed = "127.0.1.1"
+
   akka {
-    remote.server.port = 2553
+    remote.netty.port = 8553
     remote.server.message-frame-size = 16777216
     remote.client.message-frame-size = 16777216
   }
