+++ dev/src/main/resources/common.conf
@@ -1,14 +1,13 @@
 akka {
-  version = "2.0-M2"
+  version = "2.1.0"
 
   actor {
     provider = "akka.remote.RemoteActorRefProvider"
   }
 
   remote {
-    transport = "akka.remote.netty.NettyRemoteSupport"
     server {
-      hostname = "192.168.16.41"
+      hostname = "127.0.0.1"
     }
   }
 }
